tests/submission/Infraware/Forms/contents/Forms/Not Use/input_autofocus.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>input_autofocus</h3>
	</p>
	
	<li>The text control be focused when the document was loaded.</li>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  id="input_form">
	<p><input type='text' id='input_text'></p>
	</form>

	<script>
		
		var input_text = document.getElementById("input_text");
		input_text.autofocus = true;

		if (typeof(input_text.autofocus) == "boolean") {
			if (input_text.autofocus != true) {
				test(function() {
					assert_true(false, "autofocus attribute is not correct.");
				});
			}
		} else {
			test(function() {
				assert_unreached("autofocus attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
